Fix the three warnings only GCC 13 and the ASan build reported
The -Werror gate caught what the local GCC 15 build did not:

- test_decor_render.cpp: -Wformat-truncation on the `%03d` frame path.
  The two fixed 512-byte buffers become std::string/std::format, so the
  capture path cannot truncate regardless of how long the capture dir is.
- cheat_handler.cpp and pixien.cpp: -Wconversion on `x %= N` and
  `frame++ % frames`, both int-promoted results assigned back to a short.
  Explicit casts, same values.

The first is a GCC 13 diagnostic; the other two only fire in the ASan
build. Verified against a full GCC 13.4 build of the tree (0 warnings,
matching CI's 13.3) and a clean ci-asan build with -Werror on.

+++ src/interface/render/pixien.cpp
@@ -79,6 +79,6 @@ short pixieN::next_frame()
 	// frames can be 0 for a default/moved-from pixie; guard the modulo to avoid
 	// integer division-by-zero UB. set_frame already rejects out-of-range here.
 	if (frames <= 0) return 0;
-	return set_frame(frame++ % frames);
+	return set_frame(static_cast<short>(frame++ % frames));
 }
